# src/main/java/com/performance/service/dto/CalificationCalibrationProjection.java
package com.performance.service.dto;

import java.util.Date;

public interface CalificationCalibrationProjection {

    public Integer getId();
    public Integer getCalification();
    public Date getCreated();

    public String getEvaluator();
    public Integer getCalibration();
    public Date getCalibrated();
    public String getComment();
    public Integer getSuccess();

    public String getCommentcali();

    public Date getCreatedcomment();


}