# src/main/java/com/performance/service/dto/CommentCalibrationDTO.java
package com.performance.service.dto;

import lombok.*;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CommentCalibrationDTO {
    private String idsff;
    private String comment;
    private Integer idCalification;
}